Update 6
It took some time to fix up the open and save files method as I was coming across a lot of errors, but in the end the code was working properly. I also added the features of the double click to clear all text boxes ONLY when double clicking the Name Text Box, and I added function where the application will save data when the form closes.